# Rework BinaryValue