… benchmarks/DSE.Open.Benchmarks/Values/BinaryValueFromEncodedBytesBenchmarks.cs
… benchmarks/DSE.Open.Benchmarks/Values/BinaryValueFromEncodedCharsBenchmarks.cs
… benchmarks/DSE.Open.Benchmarks/Values/BinaryValueFromEncodedStringBenchmarks.cs
… benchmarks/DSE.Open.Benchmarks/Values/BinaryValueJsonBenchmarks.cs
… benchmarks/DSE.Open.Benchmarks/Values/BinaryValueToBase62Benchmarks.cs
… benchmarks/DSE.Open.Benchmarks/Values/BinaryValueToBase64Benchmarks.cs
… benchmarks/DSE.Open.Benchmarks/Values/BinaryValueToHexLowerBenchmarks.cs
… benchmarks/DSE.Open.Benchmarks/Values/BinaryValueToHexUpperBenchmarks.cs
… src/DSE.Open/HexConverter.cs
… src/DSE.Open/Text/Json/Serialization/JsonStringBinaryValueBase64Converter.cs
… test/DSE.Open.Tests/Base62ConverterTests.cs
… test/DSE.Open.Tests/BinaryValueTests.cs
… test/DSE.Open.Tests/HexConverterTests.cs