## JSON

Nice speedup and reduction in allocations for JSON round trip.

| Method         |     Mean |   Error |  StdDev | Allocated |
| -------------- | -------: | ------: | ------: | --------: |
| RoundTrip_Main | 200.9 ns | 0.34 ns | 0.32 ns |     240 B |
| RoundTrip_PR   | 158.0 ns | 0.43 ns | 0.36 ns |     112 B |

### PR

Deserialize now only allocates the number of bytes needed for the backing store.

| Method      | Mean     | Error    | StdDev   | Allocated |
|------------ |---------:|---------:|---------:|----------:|
| Serialize   | 77.32 ns | 0.204 ns | 0.181 ns |      72 B |
| Deserialize | 75.70 ns | 0.190 ns | 0.178 ns |      40 B |

## Formatting

### Base64 (default)

#### Main

| Method                |     Mean |    Error |   StdDev | Allocated |
| --------------------- | -------: | -------: | -------: | --------: |
| ToBase64String        | 14.83 ns | 0.092 ns | 0.077 ns |      64 B |
| TryFormatBase64_Chars | 16.06 ns | 0.050 ns | 0.047 ns |      64 B |

#### PR

| Method                |       Mean |     Error |    StdDev | Allocated |
| --------------------- | ---------: | --------: | --------: | --------: |
| ToBase64String        | 14.8255 ns | 0.0528 ns | 0.0494 ns |      64 B |
| TryFormatBase64_Chars |  0.1807 ns | 0.0040 ns | 0.0035 ns |         - |
| TryFormatBase64_Bytes |  9.2712 ns | 0.0221 ns | 0.0184 ns |         - |

```cs
[MemoryDiagnoser(displayGenColumns: false)]
public class BinaryValueToBase64Benchmarks
{
    private static readonly BinaryValue s_value = BinaryValue.FromString("Hello, World!");

    [Benchmark]
    public string ToBase64String() => s_value.ToBase64EncodedString();

    [Benchmark]
    public bool TryFormatBase64_Chars()
    {
        Span<char> buffer = stackalloc char[16];
        return s_value.TryFormat(buffer, out _, format: default, provider: default);
    }

    [Benchmark]
    public bool TryFormatBase64_Bytes()
    {
        Span<byte> buffer = stackalloc byte[16];
        return s_value.TryFormat(buffer, out _, format: default, provider: default);
    }
}
```

### Hex

#### Main

| Method           |     Mean |    Error |   StdDev | Allocated |
| ---------------- | -------: | -------: | -------: | --------: |
| ToHexLowerString | 26.53 ns | 0.047 ns | 0.041 ns |     160 B |
| ToHexUpperString | 11.11 ns | 0.151 ns | 0.142 ns |      80 B |

#### PR

| Method                  |     Mean |    Error |   StdDev | Allocated |
| ----------------------- | -------: | -------: | -------: | --------: |
| ToHexLowerString        | 16.59 ns | 0.291 ns | 0.273 ns |      80 B |
| ToHexUpperString        | 16.88 ns | 0.176 ns | 0.156 ns |      80 B |
|                         |          |          |          |           |
| TryFormatHexLower_Chars | 11.81 ns | 0.014 ns | 0.013 ns |         - |
| TryFormatHexUpper_Chars | 12.04 ns | 0.108 ns | 0.101 ns |         - |
|                         |          |          |          |           |
| TryFormatHexLower_Bytes | 17.94 ns | 0.034 ns | 0.030 ns |         - |
| TryFormatHexUpper_Bytes | 17.47 ns | 0.060 ns | 0.057 ns |         - |

## Creating

Creating was already calling off to the various `Convert` methods, so no perf gains when creating from encoded strings.

#### Main

| Method                  |       Mean |     Error |    StdDev | Allocated |
| ----------------------- | ---------: | --------: | --------: | --------: |
| FromBase62EncodedString | 512.566 ns | 1.6935 ns | 1.4141 ns |    1304 B |
| FromBase64EncodedString |  24.773 ns | 0.0531 ns | 0.0443 ns |      40 B |
| FromHexLowerString      |   9.380 ns | 0.0325 ns | 0.0304 ns |      40 B |
| FromHexUpperString      |   9.420 ns | 0.0220 ns | 0.0195 ns |      40 B |

#### PR

| Method                  |       Mean |     Error |    StdDev | Allocated |
| ----------------------- | ---------: | --------: | --------: | --------: |
| FromBase62EncodedString | 508.442 ns | 1.6548 ns | 1.5479 ns |    1304 B |
| FromBase64EncodedString |  24.905 ns | 0.0314 ns | 0.0262 ns |      40 B |
| FromHexLowerString      |   9.535 ns | 0.0398 ns | 0.0332 ns |      40 B |
| FromHexUpperString      |   9.573 ns | 0.0586 ns | 0.0548 ns |      40 B |

| Method                 |      Mean |    Error |   StdDev | Allocated |
| ---------------------- | --------: | -------: | -------: | --------: |
| FromBase62EncodedChars | 508.54 ns | 1.373 ns | 1.285 ns |    1304 B |
| FromBase64EncodedChars |  18.60 ns | 0.052 ns | 0.048 ns |      40 B |
| FromHexUpperChars      |  23.42 ns | 0.084 ns | 0.070 ns |      40 B |
| FromHexLowerChars      |  23.42 ns | 0.068 ns | 0.064 ns |      40 B |

| Method                 |      Mean |    Error |   StdDev | Allocated |
| ---------------------- | --------: | -------: | -------: | --------: |
| FromBase62EncodedBytes | 522.26 ns | 1.444 ns | 1.351 ns |    1368 B |
| FromBase64EncodedBytes |  15.70 ns | 0.045 ns | 0.042 ns |      40 B |
| FromHexUpperBytes      |  21.69 ns | 0.138 ns | 0.129 ns |      40 B |
| FromHexLowerBytes      |  21.67 ns | 0.043 ns | 0.040 ns |      40 B |